site/requirements.txt
django
markdown
django-simple-captcha
mysqlclient
